+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `tt rs rebootstrap`: re-bootstraps an instance.
 - `-s (--self)` flag to execute `tt` itself and don't search for other `tt`s in bin_dir
   provided in config.
+- `tt start` interactive mode with `-i` option.
 
 ### Fixed